+++ components/app/myprep/insights/MyRWInsightsStarred.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class MyPREPInsightsStarred extends React.Component {
+  render() {
+    return (
+      <div>
+        Starred
+      </div>
+    );
+  }
+}
+
+export default MyPREPInsightsStarred;