+++ CHANGELOG.md
@@ -50,7 +50,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Explore: favorite datasets
 - Explore detail: favorite dataset
-- MyRW Dataset favorites
+- MyPrep Dataset favorites
 - Explore detail tags section
 - Favorite widget functionality at Embed widget page
 - Styles for LayerInfoModal component